fix template type

+++ dScripts/newcontent/server/HalloweenManager.cpp
@@ -11,7 +11,7 @@ void HalloweenManager::OnStartup(Entity* self) {
 
 void HalloweenManager::SpawnTheBossSmashable(Entity* self) {
 
-	auto choice = GeneralUtils::GenerateRandomNumber<uint8_t>(1,3);
+	auto choice = GeneralUtils::GenerateRandomNumber<uint32_t>(1,3);
 	EntityInfo info{};
 
 	switch (choice) {
@@ -35,7 +35,7 @@ void HalloweenManager::SpawnTheBossSmashable(Entity* self) {
 			break;
 		}
 	}
-
+	
 	auto spawnedEntity = Game::entityManager->CreateEntity(info, nullptr, self);
 	Game::entityManager->ConstructEntity(spawnedEntity);
 	spawnedEntity->AddDieCallback([this, self]() {
